Rename spheropolygon filename to reflect convexity.
