+++ coxeter/shape_classes/convex_spheropolygon.py
